Apparently even normals can miss in some obj files

+++ RayTracer/ObjLoader.h
@@ -8,7 +8,7 @@
 class ObjLoader
 {
 public:
-	typedef std::vector<std::tuple<size_t, size_t, size_t>> Polygon;
+	typedef std::vector<std::tuple<size_t, long long int, long long int>> Polygon;
 
 	class Material
 	{
